# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Suraksha Setu API Documentation
  version: 1.0.0
  description: Comprehensive API documentation for Suraksha Setu disaster management platform
  contact:
    name: Suraksha Setu Team
    email: support@surakshasetu.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.surakshasetu.com
    description: Production server
tags:
  - name: Authentication
    description: User authentication and authorization endpoints
  - name: SOS
    description: Emergency SOS reporting and management
  - name: Resources
    description: Emergency resource management (hospitals, shelters, etc.)
  - name: Alerts
    description: Alert broadcasting and notification system
  - name: Dashboard
    description: Dashboard data and statistics
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session-based authentication using Passport.js (cookie)
  schemas:
    User:
      type: object
      required:
        - name
        - email
        - password
        - role
      properties:
        _id:
          type: string
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        phone:
          type: string
          pattern: '^[0-9]{10}$'
        role:
          type: string
          enum:
            - citizen
            - admin
        location:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
            address:
              type: string
        createdAt:
          type: string
          format: date-time
        isActive:
          type: boolean
          default: true
    SOS:
      type: object
      required:
        - userId
        - type
        - severity
        - description
        - location
      properties:
        _id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum:
            - flood
            - fire
            - earthquake
            - medical
            - accident
            - other
        severity:
          type: string
          enum:
            - low
            - medium
            - high
            - critical
        status:
          type: string
          enum:
            - pending
            - acknowledged
            - in-progress
            - resolved
          default: pending
        description:
          type: string
          maxLength: 500
        location:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
            address:
              type: string
        contactNumber:
          type: string
          pattern: '^[0-9]{10}$'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Resource:
      type: object
      required:
        - name
        - type
        - location
        - contact
      properties:
        _id:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            - hospital
            - shelter
            - fire-station
            - police-station
            - community-center
            - relief-camp
        location:
          type: object
          properties:
            type:
              type: string
              enum:
                - Point
            coordinates:
              type: array
              items:
                type: number
              description: '[longitude, latitude]'
            address:
              type: string
            city:
              type: string
            state:
              type: string
            pincode:
              type: string
        contact:
          type: object
          properties:
            phone:
              type: string
              pattern: '^[0-9]{10}$'
            email:
              type: string
              format: email
            website:
              type: string
              format: uri
        services:
          type: array
          items:
            type: string
        capacity:
          type: number
          minimum: 0
        currentOccupancy:
          type: number
          minimum: 0
          default: 0
        operatingHours:
          type: string
          default: 24/7
        isActive:
          type: boolean
          default: true
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
    Alert:
      type: object
      required:
        - title
        - message
        - type
        - createdBy
      properties:
        _id:
          type: string
        title:
          type: string
          maxLength: 100
        message:
          type: string
          maxLength: 500
        severity:
          type: string
          enum:
            - info
            - warning
            - danger
            - critical
          default: info
        type:
          type: string
          enum:
            - weather
            - disaster
            - health
            - security
            - general
        targetAudience:
          type: string
          enum:
            - all
            - location-based
            - admin-only
          default: all
        location:
          type: object
          properties:
            type:
              type: string
              enum:
                - Point
            coordinates:
              type: array
              items:
                type: number
            radius:
              type: number
            city:
              type: string
            state:
              type: string
        isActive:
          type: boolean
          default: true
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        readBy:
          type: array
          items:
            type: string
paths:
  /register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - passwordConfirm
                - role
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                passwordConfirm:
                  type: string
                phone:
                  type: string
                role:
                  type: string
                  enum:
                    - citizen
                    - admin
            example:
              name: John Doe
              email: john@example.com
              password: SecurePass123!
              passwordConfirm: SecurePass123!
              phone: '9876543210'
              role: citizen
      responses:
        '201':
          description: User registered
        '400':
          description: Validation error
  /login:
    post:
      summary: Log in a user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
            example:
              email: john@example.com
              password: SecurePass123!
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
  /logout:
    get:
      summary: Logout current user
      tags:
        - Authentication
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Logged out
  /profile:
    get:
      summary: Get current user's profile
      tags:
        - Authentication
      security:
        - sessionAuth: []
      responses:
        '200':
          description: User profile
  /sos:
    get:
      summary: List SOS reports
      tags:
        - SOS
      security:
        - sessionAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
          description: Filter by status
        - in: query
          name: type
          schema:
            type: string
          description: Filter by type of SOS
      responses:
        '200':
          description: List of SOS reports
    post:
      summary: Create a new SOS report
      tags:
        - SOS
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - description
                - lat
                - lng
              properties:
                type:
                  type: string
                severity:
                  type: string
                description:
                  type: string
                lat:
                  type: number
                lng:
                  type: number
            example:
              type: flood
              severity: high
              description: Water rising quickly
              lat: 28.7041
              lng: 77.1025
      responses:
        '201':
          description: SOS created
  '/sos/{id}':
    get:
      summary: Get a specific SOS report
      tags:
        - SOS
      security:
        - sessionAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SOS report
  '/sos/{id}/status':
    put:
      summary: Update SOS status (admin)
      tags:
        - SOS
      security:
        - sessionAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, acknowledged, in-progress, resolved]
      responses:
        '200':
          description: Status updated
  /api/resources:
    get:
      summary: List resources
      tags:
        - Resources
      security:
        - sessionAuth: []
      parameters:
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: lat
          schema:
            type: number
        - in: query
          name: lng
          schema:
            type: number
        - in: query
          name: radius
          schema:
            type: number
      responses:
        '200':
          description: List of resources
    post:
      summary: Create a new resource (admin)
      tags:
        - Resources
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Resource'
      responses:
        '201':
          description: Resource created
  '/api/resources/{id}':
    get:
      summary: Get a single resource
      tags:
        - Resources
      security:
        - sessionAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Resource object
    put:
      summary: Update a resource (admin)
      tags:
        - Resources
      security:
        - sessionAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Resource'
      responses:
        '200':
          description: Resource updated
    delete:
      summary: Delete a resource (admin)
      tags:
        - Resources
      security:
        - sessionAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Resource deleted
  /api/alerts:
    get:
      summary: Get alerts for user
      tags:
        - Alerts
      security:
        - sessionAuth: []
      parameters:
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: severity
          schema:
            type: string
      responses:
        '200':
          description: List of relevant alerts
    post:
      summary: Create/broadcast an alert (admin)
      tags:
        - Alerts
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Alert'
      responses:
        '201':
          description: Alert created and broadcast
  '/api/alerts/{id}/read':
    put:
      summary: Mark alert as read
      tags:
        - Alerts
      security:
        - sessionAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert marked as read
  /api/alerts/history:
    get:
      summary: Get alert history (admin)
      tags:
        - Alerts
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Alert history list
  '/api/alerts/{id}':
    delete:
      summary: Deactivate an alert (admin)
      tags:
        - Alerts
      security:
        - sessionAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert deactivated
  /dashboard:
    get:
      summary: Citizen dashboard data
      tags:
        - Dashboard
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Citizen dashboard data
  /admin/dashboard:
    get:
      summary: Admin dashboard data
      tags:
        - Dashboard
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Admin dashboard data